backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.10.4
pydantic-settings==2.1.0

# Data Processing
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# ML & Prediction
xgboost==2.0.3
shap==0.44.0
# lightgbm==4.2.0  # Commented out - CMake build issues, not needed for lifecycle detection

# Visualization & Analysis
matplotlib==3.8.2
seaborn==0.13.1

# Networking & Async
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# MongoDB (for Lifecycle Module)
motor==3.3.2
pymongo==4.6.1

# Google Trends API
pytrends==4.9.2

# Twitter/X API
tweepy==4.14.0

# Reddit API
praw==7.7.1

# Google Gemini AI
google-generativeai==0.3.2

 # Groq AI
 groq

 # PDF parsing
 pypdf==5.3.0

 # DOCX parsing
 python-docx==1.1.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
email-validator==2.3.0

# NLP (for sentiment & text analysis)
nltk==3.8.1
textblob==0.17.1

# Date & Time
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
